+++ src/App.js
@@ -6,7 +6,6 @@ import "./App.css";
 
 import Listings from "./pages/listings/Listings";
 import Form from "./pages/form/Form";
-import Form2 from "./pages/form2/Form2";
 import Login from "./pages/login/Login";
 import Ficha from "./pages/ficha/Ficha";
 import Signup from "./pages/signup/Signup";
@@ -24,12 +23,12 @@ function App() {
             <Navbar />
             <Switch>
               <Route exact path="/">
-                {!user && <Redirect to="/form2" />}
-                {user && <Form2 />}
-              </Route>
-              <Route path="/form2">
                 {!user && <Redirect to="/login" />}
-                {user && <Form2 />}
+                {user && <Form />}
+              </Route>
+              <Route path="/form">
+                {!user && <Redirect to="/login" />}
+                {user && <Form />}
               </Route>
 
               <Route path="/success">
